Update client3.py
Fixed error in getDatapoint by adding price formula
changing the value in getRatio method by modifying the ratio of stock and adding condition in where if price is = 0 return
Modifying the main Main method to print the correct values. By adding datapoints and getDataPoints so we can getRatio and print the right ratio.